// ui-frontend-portal/src/app/gambit/model/PagePartnerInfo.ts
/**
 * Gambit API
 * The Gambit API offers Endpoints for an Mobile App and a Management UI.
 *
 * OpenAPI spec version: 0.1.2
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface PagePartnerInfo {
    content?: Array<models.PartnerInfo>;

    first?: boolean;

    last?: boolean;

    number?: number;

    numberOfElements?: number;

    size?: number;

    sort?: models.Sort;

    totalElements?: number;

    totalPages?: number;

}